# swagger/user.yaml
openapi: 3.0.0
info:
  title: API Biblioteca UPEL
  version: 1.0.0
  description: Documentación de la API de gestión de usuarios y libros de la biblioteca UPEL
servers:
  - url: http://localhost:5000/api
paths:
  /users/login:
    post:
      summary: Iniciar sesión de usuario
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ci:
                  type: string
                  example: "12345678"
                password:
                  type: string
                  example: "Password123!"
      responses:
        200:
          description: Login exitoso
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    type: object
        401:
          description: Credenciales inválidas
  /users:
    post:
      summary: Registrar usuario
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ci:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                username:
                  type: string
                email:
                  type: string
                phone:
                  type: string
                user_type:
                  type: string
                  enum: [student, professor, admin]
                password:
                  type: string
      responses:
        201:
          description: Usuario creado
        400:
          description: Error de validación o email duplicado
    get:
      summary: Obtener usuarios paginados
      tags: [Users]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
          description: Número de página
        - in: query
          name: limit
          schema:
            type: integer
          description: Cantidad de usuarios por página
      responses:
        200:
          description: Lista de usuarios y paginación
  /users/{ci}:
    get:
      summary: Obtener usuario por CI
      tags: [Users]
      parameters:
        - in: path
          name: ci
          required: true
          schema:
            type: string
          description: Cédula de identidad
      responses:
        200:
          description: Usuario encontrado
        404:
          description: Usuario no encontrado
    put:
      summary: Actualizar usuario (solo admin)
      tags: [Users]
      parameters:
        - in: path
          name: ci
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                last_name:
                  type: string
                username:
                  type: string
                phone:
                  type: string
                user_type:
                  type: string
                is_active:
                  type: boolean
      responses:
        200:
          description: Usuario actualizado
        404:
          description: Usuario no encontrado
        403:
          description: Acceso denegado
    delete:
      summary: Eliminar usuario (solo admin)
      tags: [Users]
      parameters:
        - in: path
          name: ci
          required: true
          schema:
            type: string
      responses:
        200:
          description: Usuario eliminado
        404:
          description: Usuario no encontrado
        403:
          description: Acceso denegado
  /users/{ci}/restore:
    patch:
      summary: Restaurar usuario
      tags: [Users]
      parameters:
        - in: path
          name: ci
          required: true
          schema:
            type: string
      responses:
        200:
          description: Usuario restaurado
        404:
          description: Usuario no encontrado
  /users/email/{email}:
    get:
      summary: Buscar usuario por email
      tags: [Users]
      parameters:
        - in: path
          name: email
          required: true
          schema:
            type: string
      responses:
        200:
          description: Usuario encontrado
        404:
          description: Usuario no encontrado